backend/nyay-audit/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
neo4j
transformers
torch
pydantic
python-dotenv
httpx
spacy
